tf_3dunet_brats updated to new convention
