--- openfl-workspace/tf_3dunet_brats/plan/defaults/aggregator.yaml
@@ -1,4 +1,0 @@
-template : openfl.component.Aggregator
-settings :
-    db_store_rounds   : 1
-
