--- openfl-workspace/tf_3dunet_brats/plan/defaults/assigner.yaml
@@ -1,9 +1,0 @@
-template : openfl.component.RandomGroupedAssigner
-settings :
-  task_groups  :
-    - name       : train_and_validate
-      percentage : 1.0
-      tasks      :
-        - aggregated_model_validation
-        - train
-        - locally_tuned_model_validation
