--- openfl-workspace/tf_3dunet_brats/plan/defaults/collaborator.yaml
@@ -1,5 +1,0 @@
-template : openfl.component.Collaborator
-settings :
-    opt_treatment     : 'CONTINUE_LOCAL'
-    delta_updates     : True
-    db_store_rounds   : 1
