--- openfl-workspace/tf_3dunet_brats/plan/defaults/task_runner.yaml
@@ -1,1 +1,0 @@
-template: openfl.federated.task_runner.CoreTaskRunner
