--- openfl-workspace/tf_3dunet_brats/plan/defaults/tasks_fast_estimator.yaml
@@ -1,22 +1,0 @@
-aggregated_model_validation:
-  function : validate
-  kwargs   :
-    batch_size : 32
-    apply      : global
-    metrics    :
-      - accuracy
-   
-locally_tuned_model_validation:
-  function : validate
-  kwargs   :
-    batch_size : 32
-    apply      : local
-    metrics    :
-      - accuracy
-train:
-  function : train
-  kwargs   :
-    batch_size : 32
-    epochs     : 1
-    metrics    :
-    - loss
