--- openfl-workspace/tf_3dunet_brats/plan/defaults/tasks_torch.yaml
@@ -1,19 +1,0 @@
-aggregated_model_validation:
-  function : validate
-  kwargs   :
-    apply   : global
-    metrics :
-      - acc
-  
-locally_tuned_model_validation:
-  function  : validate
-  kwargs    :
-    apply: local
-    metrics :
-      - acc
-  
-train:
-  function : train_batches
-  kwargs   :
-    metrics     :
-    - loss
