+++ openfl-workspace/tf_3dunet_brats/src/taskrunner.py
@@ -3,7 +3,108 @@
 
 """You may copy this file as the starting point of your own model."""
 
+import numpy as np
 import tensorflow as tf
+
+from openfl.utilities import Metric
+from openfl.federated import TensorFlowTaskRunner
+
+
+class UNet3D(TensorFlowTaskRunner):
+    """Initialize.
+
+    Args:
+        **kwargs: Additional parameters to pass to the function
+
+    """
+
+    def __init__(self, initial_filters=16,
+                 depth=5,
+                 batch_norm=True,
+                 use_upsampling=False,
+                 **kwargs):
+        """Initialize.
+
+        Args:
+            **kwargs: Additional parameters to pass to the function
+
+        """
+        super().__init__(**kwargs)
+
+        self.model = self.create_model(
+            input_shape=self.feature_shape,
+            n_cl_out=self.data_loader.num_classes,
+            initial_filters=initial_filters,
+            use_upsampling=use_upsampling,
+            depth=depth,
+            batch_norm=batch_norm,
+            **kwargs
+        )
+        self.initialize_tensorkeys_for_functions()
+
+        self.model.summary(print_fn=self.logger.info, line_length=120)
+
+    def create_model(self,
+                     input_shape,
+                     n_cl_out=1,
+                     use_upsampling=False,
+                     dropout=0.2,
+                     print_summary=True,
+                     seed=816,
+                     depth=5,
+                     dropout_at=(2, 3),
+                     initial_filters=16,
+                     batch_norm=True,
+                     **kwargs):
+        """Create the TensorFlow 3D U-Net CNN model.
+
+        Args:
+            input_shape (list): input shape of the data
+            n_cl_out (int): Number of output classes in label (Default=1)
+            **kwargs: Additional parameters to pass to the function
+
+        """
+        
+        model = build_model(input_shape,
+                            n_cl_out=n_cl_out,
+                            use_upsampling=use_upsampling,
+                            dropout=dropout,
+                            print_summary=print_summary,
+                            seed=seed,
+                            depth=depth,
+                            dropout_at=dropout_at,
+                            initial_filters=initial_filters,
+                            batch_norm=batch_norm)
+
+        model.compile(
+            loss=dice_loss,
+            optimizer=tf.keras.optimizers.Adam(),
+            metrics=[dice_coef, soft_dice_coef],
+        )
+
+        return model
+    
+    def train_(self, batch_generator, metrics: list = None, **kwargs):
+        """Train single epoch.
+
+        Override this function for custom training.
+
+        Args:
+            batch_generator: Generator of training batches.
+                Each batch is a tuple of N train images and N train labels
+                where N is the batch size of the DataLoader of the current TaskRunner instance.
+
+            epochs: Number of epochs to train.
+            metrics: Names of metrics to save.
+        """
+        history = self.model.fit(batch_generator,
+                                 verbose=1,
+                                 **kwargs)
+        results = []
+        for metric in metrics:
+            value = np.mean([history.history[metric]])
+            results.append(Metric(name=metric, value=np.array(value)))
+        return results
 
 
 def dice_coef(target, prediction, axis=(1, 2, 3), smooth=0.0001):
